Emit correct exit code on Sass compilation errors if env != development
Use hard exit for prod, staging and testing environments (this
is especially important for CI debugging) when Sass encounters
compilation errors. On the other hand, it is useful to only emit
"soft" errors during local development; it would be frustrating
if watch/serve process would crash every time a Sass compilation
is not successful (e.g. because of a typo).