Offer preconfigured SVG optimizations via gulp-svgmin (Case 146144) (#21)
This task purposely does not strip the `xmlns` attribute, so a generous path config can be used (all SVGs independent of inline or external usage). The namespace is required for SVGs that are used via `<img src />` or CSS `background-image` but is not needed in HTML5 and may even produce issues with inline SVG used in HTML that needs to be valid XHTML. 

The idea is to optimize as much as possible automatically, but remove the namespace attributes manually on demand.