Resolve to current major version of clean-css
Use Yarn resolution in package.json to require the current
major version of clean-css (unfortunately, gulp-clean-css seems
to be dead with the pertinent PR remaining unmerged since 2021).

clean-css 5.3.2 includes a fix that enables us to use conainer
queries without losing them to CSS minification.

We no longer need to disable URL rebasing as the new default
is `false`.

This seems to be a good time to bump compatibility up to ie11 (the
default is >11 which is still a bit too strict for us in 2023).