Preserve hashes when rebasing urls (#45)
Fixes two bugs when working with hashes in `url()` values of properties
in CSS:

1. Adding the hash to the final path string by default. This is safe,
because the hash property of the asset object is either the hash itself
or an empty string. This makes working with hashes possible in the first
place.
2. When referencing inlined resources by a hash, the default path string
is no longer added, which resulted in not beeing able to reference
inline resources at all.