Support TypeScript files as bundle sources (Case 199344) (#47)
Adds support for TypeScript files to the Webpack bundling task.

The gulp task including the modifications from this PR is already in use
in
https://github.com/webfactory/webfactory-de-projekt-controlling/blob/3a61cad789bc893c8fd406b279a0d4bd57b86e90/webpack-ts.js,
which can be set back to using the upstream task after this has been
merged.

Apart from being necessary for our Projekt Controlling app, I'd like to
make it easier to use TypeScript in webfactory projects to encourage its
use.

It might be "cleaner" to open the gulp preset for downstream Webpack
config modifications rather than adding TypeScript upstream. However,
this is already how we decided to handle Svelte support, which is used
by a similarly small amount of webfactory projects, and it seems to be
easier right now – the downsides of requiring TypeScript upstream should
not be this big, given this is just a dev dependency.

That doesn't mean I wouldn't be positive designing an interface for
additional webpack configuration from downstream, and maybe bringing the
syntax expected in `gulp-config.js` closer to the well documented
Webpack config format.